ssh-client/requirements.txt
aiohttp==3.8.1
python-socketio==5.5.2
requests==2.28.1
